[ASV-1738] Refactored analytics by abstracting all dependencies on WalletInstallAnalytics
